% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdata.ex.R
\docType{data}
\name{cdata.ex}
\alias{cdata.ex}
\title{Clustered dataset for centering example}
\format{
A wide data frame of 60 observations. Used for discussing within and between group effects.
\describe{
  \item{x}{The predictor.}
  \item{y}{The outcome of interest.}

}
}
\usage{
cdata.ex
}
\description{
Dataset of 60 observations from 3 clusters.
}
\keyword{datasets}
