% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_CR2.R
\name{glance.CR2}
\alias{glance.CR2}
\title{Glance at goodness-of-fit statistics}
\usage{
\method{glance}{CR2}(x, ...)
}
\arguments{
\item{x}{A `CR2` object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
\code{glance} returns one row with the columns:
  \item{nobs}{the number of observations}
  \item{sigma}{the square root of the estimated residual variance}
  \item{logLik}{the data's log-likelihood under the model}
  \item{AIC}{Akaike Information Criterion}
  \item{BIC}{Bayesian Information Criterion}
  \item{r2.marginal}{marginal R2 based on fixed effects only using method of Nakagawa and Schielzeth (2013)}
  \item{r2.conditional}{conditional R2 based on fixed and random effects using method of Nakagawa and Schielzeth (2013)}
}
\description{
Helper function used to obtain supporting fit statistics for multilevel models. The R2s are computed using the `performance` package.
}
