% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mean.R
\name{group_mean}
\alias{group_mean}
\title{Computes the group mean of a variable}
\usage{
group_mean(x, grp)
}
\arguments{
\item{x}{Variable to compute the mean for (e.g., dataframe$varname)}

\item{grp}{Cluster/grouping variable (e.g., dataframe$cluster)}
}
\description{
Computes the group mean of a variable
}
