% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupcenter.R
\name{groupcenter}
\alias{groupcenter}
\title{Group-mean center a variable}
\usage{
groupcenter(var, grp)
}
\arguments{
\item{var}{Variable to center (e.g., dataframe$varname)}

\item{grp}{Cluster/grouping variable (e.g., dataframe$cluster)}
}
\value{

}
\description{
Also referred to as centering within cluster.
}
\examples{
data(mtcars)
mtcars$mpg.gpc <- groupcenter(mtcars$mpg, mtcars$cyl)
}
