% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Htest.R
\name{Htest}
\alias{Htest}
\title{Test for homoskedasticity at level one}
\usage{
Htest(newdata, fml, group)
}
\arguments{
\item{newdata}{data to be used}

\item{fml}{level 1 formula}

\item{group}{grouping variable (in quotes)

Based on Raudenbush and Bryk (2002).
A statistically significant Chisq indicates heteroskedasticity.
Output shows the H statistic, degrees of freedom, and p value.}
}
\description{
Test for homoskedasticity at level one
}
