% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mixed.R, R/robust_mixed_2.R
\name{MatSqrtInverse}
\alias{MatSqrtInverse}
\title{Compute the inverse square root of a matrix}
\usage{
MatSqrtInverse(A)

MatSqrtInverse(A)
}
\arguments{
\item{A}{A matrix.}
}
\description{
Compute the inverse square root of a matrix

Compute the inverse square root of a matrix
}
