% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmiss.R
\name{nmiss}
\alias{nmiss}
\title{Amount of missing data per variable}
\usage{
nmiss(dat)
}
\arguments{
\item{dat}{Data frame that you want to inspect}
}
\description{
Amount of missing data per variable
}
\examples{
data(mtcars)
mtcars[c(2:3), 4] <- NA #create NAs
nmiss(mtcars)

}
