% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_test1.R
\docType{data}
\name{ri_test1}
\alias{ri_test1}
\title{Sample dataset 1 for testing the likelihood ratio test}
\format{
A data frame with 900 observations from 30 groups and 4 variables:
\describe{
  \item{y}{The outcome variable.}
  \item{w1}{A level-2 predictor.}
  \item{x1}{A level-1 predictor}
  \item{group}{The cluster identifier}
}
}
\usage{
data(ri_test1)
}
\description{
Example data for testing the need for a random intercept. Illustrates the need to adjust the p values for a modified LRT.
}
\keyword{datasets}
