% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_test2.R
\docType{data}
\name{ri_test2}
\alias{ri_test2}
\title{Sample dataset 2 for testing the likelihood ratio test (LRT)}
\format{
A data frame with 3,000 observations from 30 groups and 4 variables:
\describe{
  \item{y}{The outcome variable.}
  \item{w1}{A level-2 predictor.}
  \item{x1}{A level-1 predictor}
  \item{group}{The cluster identifier}
}
}
\usage{
data(ri_test2)
}
\description{
Example data for testing the need for a random intercept. LRT results show that a random slope is not warranted.
}
\keyword{datasets}
