% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mixed.R
\name{robust_mixed}
\alias{robust_mixed}
\title{Robust standard errors for mixed models}
\usage{
robust_mixed(m1, digits = 4, Gname = NULL)
}
\arguments{
\item{m1}{lme4 or nlme model object}

\item{digits}{Number of digits for output}

\item{Gname}{Group/cluster name if more than two levels of clustering}
}
\description{
If there are more than two levels of clustering, the clustering variable should
be set at the highest level. Should not be used for inferential statistical testing purposes
if there are only a few clusters (e.g., < 40). The robust standard errors (CR0) are
based on the formulation of Liang and Zeger (1986). For a few clusters, use the
CR2 version (using the clubSandwich package).
}
\examples{
data(mtcars)
require(lme4)
robust_mixed(lmer(mpg ~ wt + am + (1|cyl), data = mtcars))
}
\references{
Liang, K.Y., & Zeger, S. L. (1986). Longitudinal data analysis using generalized linear models. Biometrika, 73(1), 13–22.
}
