% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mixed.R, R/robust_mixed_2.R
\name{satdf}
\alias{satdf}
\title{For computing Satterthwaite degrees of freedom}
\usage{
satdf(mod)

satdf(mod)
}
\arguments{
\item{mod}{An lme, lmerMod, or lmerModLmerTest object.}
}
\description{
For computing Satterthwaite degrees of freedom

For computing Satterthwaite degrees of freedom
}
