% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{summary.mixPV}
\alias{summary.mixPV}
\title{Create summary output from the mixPV function}
\usage{
\method{summary}{mixPV}(object, dfadj = TRUE, ...)
}
\arguments{
\item{object}{The mixPV object}

\item{dfadj}{If set to TRUE (default), uses newer df computation. If FALSE, uses standard Rubin pooling formula.}

\item{...}{Additional unspecified options.}
}
\description{
Create summary output from the mixPV function
}
