% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspend.R
\docType{data}
\name{suspend}
\alias{suspend}
\title{Suspension data from Virginia}
\format{
Dataset:
\describe{
  \item{school}{School identifier}
  \item{pminor}{Percent minority enrollment at school}
  \item{male}{1 = male, 0 = female}
  \item{sus}{Whether the student was suspended (1 = yes) in the
  school year or not (0 = no). Self reported.}
  \item{fight}{If the student got into one or more fights (1 = yes) in
  the school year}
  \item{gpa}{Students self-reported GPA; 1 = D to 4 = A}

}
}
\usage{
data(suspend)
}
\description{
Data from 8465 students from 100 schools in
Virginia
}
\keyword{datasets}
