% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thai.sm.R
\docType{data}
\name{thai.sm}
\alias{thai.sm}
\title{Thai data from PISA (reduced)}
\format{
A data frame with 4271 rows and 7 variables:
\describe{
  \item{math}{First plausible value in mathematics.}
  \item{escs}{Index of economic, social, and cultural status.}
  \item{intmat}{Mathematics interest.}
  \item{schoolid}{School identifier}
  \item{othl}{Spoke another language at home other than Thai. 1 = yes, 0 = no.}
  \item{private}{Private school. 1 = yes, 0 = no.}
  \item{schsize}{Total school enrolment.}
}
}
\source{
\url{https://gpseducation.oecd.org/CountryProfile?primaryCountry=THA}
}
\usage{
data(thai)
}
\description{
Example data to be used for centering
}
\keyword{datasets}
