% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tidy.mixPV}
\alias{tidy.mixPV}
\title{Helper function to allow use with modelsummary}
\usage{
\method{tidy}{mixPV}(x, dfadj = TRUE, ...)
}
\arguments{
\item{x}{The mixPV model object.}

\item{dfadj}{If set to TRUE (default), uses newer df computation. If FALSE, uses standard Rubin pooling formula.}

\item{...}{Additional unspecified options.}
}
\description{
Helper function to allow use with modelsummary
}
