% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide.R
\docType{data}
\name{wide}
\alias{wide}
\title{Wide dataset to be used for growth modeling}
\format{
A wide data frame of 30 observations:
\describe{
  \item{studentid}{Factor indicating student identification}
  \item{int}{treatment or control}
  \item{female}{1 = female, 0 = male}
  \item{fall_k}{Reading scores in fall kindergarten}
  \item{spring_k}{Reading scores in spring kindergarten}
  \item{spring_g1}{Reading scores in spring first grade}

}
}
\usage{
wide
}
\description{
A dataset containing 30 observations with reading scores
taken in the fall kindergarten, spring kindergarten, and
spring first grade
}
\keyword{datasets}
