% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmiss.R
\name{nmiss}
\alias{nmiss}
\title{Amount of missing data per variable}
\usage{
nmiss(dat)
}
\arguments{
\item{dat}{Data frame that you want to inspect.}
}
\value{
By default, this function will print the following items to the console
\itemize{
\item The percent of missing data per variable.
\item The percent of complete cases (range: 0 to 1).
\item Suggested number of datasets to impute when using multiple imputation.
}
}
\description{
Amount of missing data per variable
}
\examples{
data(mtcars)
mtcars[c(2:3), 4] <- NA #create NAs
nmiss(mtcars)

}
