% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_mixed.R
\name{robust_mixed}
\alias{robust_mixed}
\title{Robust standard errors for mixed models}
\usage{
robust_mixed(m1, digits = 4, satt = TRUE, Gname = NULL)
}
\arguments{
\item{m1}{lme4 or nlme model object}

\item{digits}{Number of digits for output}

\item{satt}{Satterthwaite degrees of freedom approximation; TRUE or FALSE}

\item{Gname}{Group/cluster name if more than two levels of clustering

If there are more than two levels of clustering, the clustering variable should
set at the highest level}
}
\value{

}
\description{
Robust standard errors for mixed models
}
\examples{
data(mtcars)
require(lmer)
robust_mixed(lmer(mpg ~ wt + am + (1|cyl), data = mtcars))
}
