% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satdf.R
\name{satdf}
\alias{satdf}
\title{Compute Satterthwaite degrees of freedom}
\usage{
satdf(m1, type = "none", Vinv2, Vm2, br2, Gname = NULL)
}
\arguments{
\item{m1}{The \code{lmerMod} or \code{lme} model object.}

\item{type}{The type of cluster robust correction used (i.e., CR2 or none).}

\item{Vinv2}{Inverse of the variance matrix.}

\item{Vm2}{The variance matrix.}

\item{br2}{The bread component.}

\item{Gname}{The group (clustering variable) name'}
}
\description{
Function to compute empirical degrees of freedom
based on Bell and McCaffrey (2002).
}
\author{
Francis Huang, \email{huangf@missouri.edu}

Bixi Zhang, \email{bixizhang@missouri.edu}
}
